moved auth related files together
