+++ package.json
@@ -15,7 +15,7 @@
     "pretest": "bunx tsc --noEmit",
     "server": "bun ./examples/node/http.ts 3000",
     "test": "bun test",
-    "bench": "bun ./src/*.bench.ts",
+    "bench": "bun ./src/**/*.bench.ts",
     "build": "bun build --compile ./index.ts --outfile substreams-sink-webhook",
     "format": "biome format --write . && biome check --apply .",
     "lint": "biome check ."
