+++ src/auth/signer.ts
